Common model code moved to approval controller
